[Agent] modify netns id u32 to u64

+++ agent/src/flow_generator/protocol_logs/mod.rs
@@ -223,9 +223,9 @@ pub struct AppProtoLogsBaseInfo {
     #[serde(skip)]
     pub is_vip_interface_dst: bool,
     #[serde(skip)]
-    pub netns_id_0: u64,
+    pub netns_id_0: u32,
     #[serde(skip)]
-    pub netns_id_1: u64,
+    pub netns_id_1: u32,
 }
 
 pub fn duration_to_micros<S>(d: &Duration, serializer: S) -> Result<S::Ok, S::Error>
@@ -305,8 +305,8 @@ impl From<AppProtoLogsBaseInfo> for flow_log::AppProtoLogsBaseInfo {
             syscall_cap_seq_1: f.syscall_cap_seq_1 as u32,
             gpid_0: f.gpid_0,
             gpid_1: f.gpid_1,
-            netns_id_0: f.netns_id_0 as u32,
-            netns_id_1: f.netns_id_1 as u32,
+            netns_id_0: f.netns_id_0,
+            netns_id_1: f.netns_id_1,
         }
     }
 }
